@@ -1,13 +1,22 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
+import './index.css';
+import {
+    BrowserRouter,
+    Routes,
+    Route} from "react-router-dom";
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import MainPage from "./routes/MainPage";
+import EditPage from "./routes/EditPage";
 
-ReactDOM.render(
-    <React.StrictMode>
-        <App />
-    </React.StrictMode>,
-  document.getElementById('root')
+const root = ReactDOM.createRoot(
+  document.getElementById('root') as HTMLElement
+);
+root.render(
+  <React.StrictMode>
+      <App />
+  </React.StrictMode>
 );
 
 // If you want to start measuring performance in your app, pass a function
